# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Alkemio OIDC Challenge Service
  version: 0.1.0
  description: >-
    External service that processes Hydra login and consent challenges, retrieves
    identity traits from Kratos, and redirects Synapse clients with preserved
    semantics.
servers:
  - url: https://oidc-service.local
paths:
  /v1/oidc/login:
    get:
      summary: Resolve Hydra login challenge
      description: Handles Hydra login challenge callbacks and returns a redirect to Synapse.
      operationId: getLoginChallenge
      security:
        - HydraLoginChallenge: []
      parameters:
        - name: login_challenge
          in: query
          required: true
          schema:
            type: string
            description: Hydra-generated login challenge identifier.
        - name: request_id
          in: header
          required: false
          schema:
            type: string
            description: Correlation ID propagated from upstream proxies.
      responses:
        "302":
          description: Redirect to Synapse with Hydra acceptance response.
          headers:
            Location:
              description: Final redirect URL back to Synapse.
              schema:
                type: string
        "400":
          $ref: '#/components/responses/ClientError'
        "404":
          $ref: '#/components/responses/InvalidChallenge'
        "503":
          $ref: '#/components/responses/Maintenance'
  /v1/oidc/consent:
    get:
      summary: Resolve Hydra consent challenge
      description: Handles Hydra consent challenge callbacks and authorises the session.
      operationId: getConsentChallenge
      security:
        - HydraConsentChallenge: []
      parameters:
        - name: consent_challenge
          in: query
          required: true
          schema:
            type: string
            description: Hydra-generated consent challenge identifier.
        - name: request_id
          in: header
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to Synapse with Hydra consent acceptance response.
          headers:
            Location:
              description: Final redirect URL back to Synapse.
              schema:
                type: string
        "400":
          $ref: '#/components/responses/ClientError'
        "404":
          $ref: '#/components/responses/InvalidChallenge'
        "503":
          $ref: '#/components/responses/Maintenance'
  /health/live:
    get:
      summary: Liveness check
      operationId: getLiveness
      security: []
      responses:
        "200":
          description: Service process is running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LivenessState'
  /health/ready:
    get:
      summary: Readiness check
      operationId: getReadiness
      security: []
      responses:
        "200":
          description: Service is ready to accept challenges.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessState'
        "503":
          description: Service cannot accept challenges.
          headers:
            Retry-After:
              description: Seconds until maintenance is expected to lift (when maintenance is active).
              schema:
                type: integer
                minimum: 0
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessState'
  /metrics:
    get:
      summary: Prometheus metrics endpoint
      operationId: getMetrics
      security: []
      responses:
        "200":
          description: Text-based metrics exposition format.
          content:
            text/plain:
              schema:
                type: string
components:
  securitySchemes:
    HydraLoginChallenge:
      type: apiKey
      in: query
      name: login_challenge
      description: Hydra-issued login challenge token attached to callback requests.
    HydraConsentChallenge:
      type: apiKey
      in: query
      name: consent_challenge
      description: Hydra-issued consent challenge token attached to callback requests.
  schemas:
    LivenessState:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [alive]
          description: Indicates the process is running.
    ErrorPayload:
      type: object
      required:
        - error
        - message
        - challengeId
      properties:
        error:
          type: string
          description: Machine-friendly error code.
        message:
          type: string
          description: Human-readable error summary.
        challengeId:
          type: string
          description: Hydra challenge identifier tied to the failure.
        missingTraits:
          type: array
          description: Traits that caused a 400 validation failure.
          items:
            type: string
        timestamp:
          type: string
          format: date-time
          description: UTC timestamp of the error.
    ReadinessState:
      type: object
      required:
        - status
        - hydra
        - kratos
      properties:
        status:
          type: string
          enum: [ready, initializing, maintenance, degraded]
        hydra:
          type: string
          enum: [ok, unreachable, unauthorized]
        kratos:
          type: string
          enum: [ok, unreachable, unauthorized]
        maintenance:
          type: boolean
          description: Indicates the maintenance toggle state.
        version:
          type: string
  responses:
    ClientError:
      description: Validation or trait mapping failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    InvalidChallenge:
      description: Hydra challenge not found or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    Maintenance:
      description: Service is in operator maintenance mode.
      headers:
        Retry-After:
          description: Seconds until maintenance is expected to lift.
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            type: object
            required:
              - error
              - message
            properties:
              error:
                type: string
                example: maintenance_mode
              message:
                type: string
                example: OIDC service temporarily disabled for maintenance.